Better orphan handling when finding new versions during expansion.
When new version of existing lib is found, consider the paths of the
prior selected version to now be omitted paths. Find any libs whose
selected paths are covered by those now omitted paths and deselect them.
That is, if a lib was selected only as a transitive dep of a now
unselected version, unselect it too.

When checking for whether it's a new lib, consider both no lib and no
selection to be the same case.

This covers most of the work the previous cut-orphans was doing, so it
now just cuts libs with no selected version.