Update Maven core deps to 3.8.2.
The 3.8.2 update includes a breaking regression that incorrectly returns
absolute paths from the pom model (for the src and resource dirs). This
seems to come from the change in https://issues.apache.org/jira/browse/MNG-7170.
The code changes work around this case and handle the absolute path as
before. Bug filed for this at:
https://issues.apache.org/jira/browse/MNG-77218.